AString::EditDistanceI() has its own header/source LevenshteinDistance.h/cpp
